Add files via upload
Hi there, I finally able to write it's full code.I'm new and thus i faced alot of difficulties in it. 
It was very difficult for me to understand the complexities of flex box but somehow i did it.
I know that there are some problems in the code all i can say is to let me know how to fix those :).